+++ packages/nodejs/src/client.ts
@@ -206,7 +206,7 @@ export default class ConcordiumNodeClient {
      * @returns the account info for the provided account address, undefined is the account does not exist
      */
     async getAccountInfo(
-        accountAddress: Address | CredentialRegistrationId,
+        accountAddress: string | Address | CredentialRegistrationId,
         blockHash: string
     ): Promise<AccountInfo | undefined> {
         if (!isValidHash(blockHash)) {
@@ -214,11 +214,17 @@ export default class ConcordiumNodeClient {
         }
 
         const getAddressInfoRequest = new GetAddressInfoRequest();
-        if (accountAddress instanceof Address) {
+
+        if (typeof accountAddress === 'string') {
+            getAddressInfoRequest.setAddress(accountAddress);
+        } else if ('address' in accountAddress) {
             getAddressInfoRequest.setAddress(accountAddress.address);
-        } else {
+        } else if ('credId' in accountAddress) {
             getAddressInfoRequest.setAddress(accountAddress.credId);
+        } else {
+            throw new Error('Invalid accountAddress input');
         }
+
         getAddressInfoRequest.setBlockHash(blockHash);
 
         const response = await this.sendRequest(