+++ packages/common/CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Additional arguments to the JSON-RPC HttpProvider, to enable is to receive and forward cookies.
 
+### Fixed
+
+- getAccountInfo no longer relies on instanceof to determine the type of input.
+
 ## 5.0.0 2022-9-29
 
 ### Added